Use constexpr if to select correct types for MPI
Updated getMpiType to work at compiletime. This function was already
likely to be inlined, but the big advantage here is that if you use an
incorrect type you get a compiler error not a runtime error. Since
this is now a constexpr function I removed the timer block because
without it the function will almost certainly be run at compile time and
inline.