classPtn: add gather function and test
Defining the ownership of all geometric model entities for a distributed
mesh can result in some ranks only having part of ownership info.  The
Gather function collects the ownership info on rank 0.  During the
rendezvous setup phase rank 0 will write the ownership info to ADIOS2
for the non-rendezvous ranks to read.

In the current implementation of wdmapp_coupling repo the
test_ohClassPtn[_appRibPtn].cpp builds the ownership infomation based on
the ownership of the mesh entities.