Initialize client Partition type based on server partition
The redev client should not need to know what type of partition scheme is being used, or initialize that partition. This modification initializes the client partition type based on the partition type on the server. This is the first step to a safer interface where the client cannot attempt to set a partition.